requirements.txt
echo "fastapi
uvicorn
sqlalchemy
jinja2
python-dotenv
requests
google-api-python-client
youtube-transcript-api
anthropic
python-multipart" > requirements.txt